Shyamala v3 changes (#1)
feat: analytics engine v3 feature changes

+++ modules/ibm-analytics-engine-api/src/main/java/com/ibm/cloud/iaesdk/ibm_analytics_engine_api/v2/model/AnalyticsEngineClusterNode.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright IBM Corp. 2020.
+ * (C) Copyright IBM Corp. 2021.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -22,7 +22,7 @@ import com.ibm.cloud.sdk.core.service.model.GenericModel;
  */
 public class AnalyticsEngineClusterNode extends GenericModel {
 
-  protected Double id;
+  protected Long id;
   protected String fqdn;
   protected String type;
   protected String state;
@@ -42,7 +42,7 @@ public class AnalyticsEngineClusterNode extends GenericModel {
    *
    * @return the id
    */
-  public Double getId() {
+  public Long getId() {
     return id;
   }
